Rearrange Kalibro Configurations' tests
 * Now each class has its own test file
